requirements.txt
flask-cors==2.0.1
requests==2.3.0
psycopg2==2.6.1
